+++ client/src/components/Footer/index.js
@@ -2,7 +2,11 @@ import React from "react";
 import "./style.css";
 
 function Footer(props) {
-    return <div className="footer"> Here is our footer</div>;
+    return <footer class="footer clearfix">
+    <div>
+                <p>Made with coffee in Austin  &copy;2019</p>	
+    </div>
+</footer>
   }
   
   export default Footer;
