+++ client/src/components/Footer/style.css
@@ -2,6 +2,9 @@
     position: fixed;
     bottom: 0px;
     margin-top: 10px;
-    background-color: aquamarine;
+    background-color: rgb(11, 56, 2);
     width: 100%;
+    color: white;
+    height: 20px;
+    font-size: 15px;
 }
