+++ client/src/components/Login/index.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import StyledFirebaseAuth from 'react-firebaseui/StyledFirebaseAuth';
 import firebase from 'firebase';
+import "./style.css";
  
 // Configure Firebase.
 const config = {
@@ -55,7 +56,8 @@ class Login extends React.Component {
           <span>
             {/* <div>Signed In!</div>
             <button onClick={() => this.signOut()}>Sign Out!</button> */}
-            <h1>Welcome {firebase.auth().currentUser.displayName}. Our story is your story. </h1>
+            <h1 className="heading">Welcome {firebase.auth().currentUser.displayName}.</h1>
+            {/* <h2 className="heading2">Our story is your story.</h2> */}
           </span>
         ) : (
           <StyledFirebaseAuth uiConfig={this.uiConfig} firebaseAuth={firebase.auth()} />
