+++ client/src/components/SignOut/index.js
@@ -4,14 +4,13 @@ import Login from "../Login";
 
 function SignOut(props) {
   return (
-    <Login>
       <div className="SignOut">
         <div>
           <div>Signed In!</div>
           <button onClick={() => this.signOut()}>Sign Out!</button>
         </div>
       </div>
-    </Login>
+
   )
 }
