+++ client/src/components/TextBox/index.js
@@ -4,8 +4,10 @@ import "./style.css";
 const TextBox = (props) => {
     return (
     <div className="textBoxDiv">
-    <p>Please add your sentance here</p>
+    <p className="direction">Please add your sentance here</p>
     <input className="textbox" type="text"/>
+    
+    <div><button className="submit">Submit</button></div>
     </div>
     )
 }
